+++ resources/views/examples/stack/aligned.blade.php
@@ -1,0 +1,4 @@
+<x-ui.stack direction="row" justify="between" align="center" class="rounded-lg border p-4">
+    <div class="text-sm font-medium">Project status</div>
+    <x-ui.badge>Active</x-ui.badge>
+</x-ui.stack>